No longer serializing file as pbuf blocks.
This should be faster and a little easier to deal with. Right now we just
get a header telling us how big the file is and then pass that off to a
simple io.Copy

Note that there's some quirky stuff in there for preventing the
ReadLengthPrefixedCollection code from dipping into the file itself.